refactor: s3 업로드 관련 코드 리팩토링
