+++ app/aws.py
@@ -1,0 +1,10 @@
+import boto3
+
+from app.core.config import settings
+
+client_s3 = boto3.client(
+        's3',
+        aws_access_key_id=settings.AWS_ACCESS_KEY,
+        aws_secret_access_key=settings.AWS_SECRET_KEY,
+        region_name=settings.AWS_REGION
+    )
